revert - re-ordered functions
